+++ apispec/ext/marshmallow/__init__.py
@@ -17,7 +17,7 @@ NAME = 'apispec.ext.marshmallow'
 
 def schema_definition_helper(spec, name, schema, **kwargs):
     """Definition helper that allows using a marshmallow
-    :class:`Schema <marshmallow.Schema>` to provide Swagger
+    :class:`Schema <marshmallow.Schema>` to provide OpenAPI
     metadata.
 
     :param type schema: A marshmallow Schema class.